feat: adding requirements.txt from development branch for testing purposes
chore: changing python version specified in python_ci.ym

feat: adding requirements.txt to branch for pip installation and gh actions

fix: adding required dependencies for CI
