+++ .github/workflows/python-ci.yml
@@ -10,12 +10,13 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
-      - name: Set up Python 3.9
+      - name: Set up Python 3.8
         uses: actions/setup-python@v2
         with:
-          python-version: 3.9
+          python-version: 3.8
       - name: Install dependencies
         run: |
+          sudo apt-get install libsndfile1
           python -m pip install --upgrade pip
           pip install flake8 pytest
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
@@ -25,6 +26,9 @@ jobs:
           flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
           # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+      - name: Set pythonpath
+        run: |
+          echo "PYTHONPATH=/home/runner/work/speech_verification_public/speech_verification_public/" >> $GITHUB_ENV
       - name: Test with pytest
         run: |
-          pytest
+          pytest speaker_verification/
